Converted 2d vectors to use double
We get better precision, and since we are not using these in interleaved
arrays we can use double

+++ src/types2d.cpp
@@ -1,4 +1,4 @@
 #include "types2d.hpp"
 
 // Check that the compiler did not pad any of the  structs
-static_assert(sizeof(GE2Dvector) == sizeof(GLfloat)*2, "Struct GE2DVector is padded ");
+static_assert(sizeof(GE2Dvector) == sizeof(GLdouble)*2, "Struct GE2DVector is padded ");
